[Validate RAFlow] Remove max value from the integer number

+++ rtags/validate_number.go
@@ -46,7 +46,8 @@ type FloatNumberValidator struct {
 // Validate method for IntegerNumberValidator
 func (v IntegerNumberValidator) Validate(val interface{}) error {
 	// don't panic
-	n, ok := val.(int) // TODO(Akshay): Convert it to int64 and than check its type
+	n, ok := val.(int64)
+
 	if !ok {
 		return fmt.Errorf("should be type of number")
 	}
